+++ src/resqml2_0_1/Activity.h
@@ -83,7 +83,7 @@ namespace RESQML2_0_1_NS
 		 *
 		 * @returns	The parameter count.
 		 */
-		DLL_IMPORT_OR_EXPORT unsigned int getParameterCount() const final;
+		DLL_IMPORT_OR_EXPORT uint64_t getParameterCount() const final;
 
 		/**
 		 * Get the count of all the parameters which have got the same title.
@@ -92,7 +92,7 @@ namespace RESQML2_0_1_NS
 		 *
 		 * @returns	The parameter count.
 		 */
-		DLL_IMPORT_OR_EXPORT unsigned int getParameterCount(const std::string & paramTitle) const final;
+		DLL_IMPORT_OR_EXPORT uint64_t getParameterCount(const std::string & paramTitle) const final;
 
 		/**
 		 * Gets parameter title
@@ -191,7 +191,7 @@ namespace RESQML2_0_1_NS
 		 *
 		 * @returns	The integer quantity parameter value.
 		 */
-		DLL_IMPORT_OR_EXPORT std::vector<int64_t> getIntegerQuantityParameterValue(const std::string & paramTitle) const final;
+		DLL_IMPORT_OR_EXPORT std::vector<int32_t> getIntegerQuantityParameterValue(const std::string & paramTitle) const final;
 
 		/**
 		 * Gets integer quantity parameter value
